launcher/requirements.txt
inquirer ~= 3.4
PyYAML ~= 6.0